git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@4894 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/arch/i386/pmap.h
@@ -37,7 +37,7 @@ struct Pagemap {
     ptent_t pm_ent[NPTENTRIES];
 };
 
-void page_init(uint32_t lower_kb, uint32_t upper_kb);
+void page_init(uint64_t lower_kb, uint64_t upper_kb);
 
 #endif /* !__ASSEMBLER__ && JOS_KERNEL */
